+++ src/utilities/gestor_de_hooks.js
@@ -1,0 +1,3 @@
+module.exports = function (deployer) {
+    return deployer.dependencies.generar_gestor_de_hooks();
+};