+++ src/www/files/prueba.js
@@ -1,0 +1,4 @@
+
+const x = Castelog.metodos.un_hook_global("al_x",function( parametros ) {
+},{ 
+});